Add support to callables instead of plain functions and allow changing the filters too.
With this, it is possible to use a callable (i.e. an array containing the name of the class and the name of the method in that class to use), both for singular and plural translations. Now a couple of filters ("trans" and "transchoice") are set, so that the interface homogeneous and matches the one provided by symfony2.